Take into account gitignore rules defined in parent directories
This change introduces [gitignore.View].

View represents a view on a directory tree that takes into account
all applicable .gitignore files. The directory tree does NOT need
to be the repository root.

For example: with a repository root at "myrepo", a view can be
anchored at "myrepo/someproject" and still respect the ignore
rules defined at "myrepo/.gitignore".

We use this functionality to synchronize files from a path nested
in a repository while respecting the repository's ignore rules.