Reload .gitignore files if they have changed
This commit changes the code in repository.go to lazily load gitignore files
as opposed to the previous eager approach. This means that the signature of
the `Ignore` function family has changed to return `(bool, error)`.

The fileset wrapper under git/ internally taints all gitignore objects to
force a call to [os.Stat] followed by a reload if they have changed.