Use filer in sync command
This deprecates usage of the `repofiles` package in favor
of the filer package and consolidates the code paths into WSFS.

Note: one potentially breaking change here is the following.
If a file at `foo/bar.txt` is created and removed, the directory
`foo` is kept around because we do not perform directory tracking.
If subsequently we need to write a file at `foo`, it will result
in an `fs.ErrExist` because it is impossible to overwrite a directory.

The previous implementation performed a recursive delete of the path
if this happened, where this implementation will return the `fs.ErrExist`
error to the user.

We can mitigate this in one of two ways:
* Track directories to remove as part of a `diff` and remove them
* Attempt to remove an empty directory tree if we see this error
* ...?