Enable environment overrides for job clusters
While they are a slice, we can identify a job cluster by its job cluster key.
A job definition with multiple job clusters with the same key is always
invalid.  We can therefore merge definitions with the same key into one.
This is compatible with how environment overrides are applied; merging a slice
means appending to it. The override will end up in the job cluster slice of the
original, which gives us a deterministic way to merge them.