Prompt once for a client profile
The previous implementation was prone to infinite looping for the account
client due to a mismatch in determining what constitutes an account client.

Ultimately, this code must never infinite loop. If a user is prompted and
selects a profile that cannot be used, they should receive that feedback
immediately and try again, instead of being prompted again.