Increase timeout waiting for job run to 1 day
It's not uncommon for job runs to take more than 2 hours. On the client side we
should not stop waiting for a job to complete if it is intentionally running
for a long time. If a job isn't supposed to run this long, the user can specify
a run timeout in the job specification itself.