Fix rendering of streaming response
The update to the Go SDK v0.23.0 in #772 included a change to make the billable
usage API return its streaming response. This still did not make the command
print out the CSV returned by the API, however. To do so, we call
`cmdio.RenderReader` in case the response is a byte stream.

Fixes #574.