Resolve configuration before performing verification
If a bundle configuration specifies a workspace host, and the user specifies a
profile to use, we perform a check to confirm that the workspace host in the
bundle configuration and the workspace host from the profile are identical. If
they are not, we return an error. The check was introduced in #571.

Previously, the code included an assumption that the client configuration was
already loaded from the environment prior to performing the check. This was not
the case, and as such if the user intended to use a non-default path to
`.databrickscfg`, this path was not used when performing the check.

The fix does the following:
* Resolve the configuration prior to performing the check.
* Don't treat the configuration file not existing as an error.
* Add unit tests.

Fixes #884.