Add configuration normalization code
This is similar to #904 but instead of converting the dynamic configuration to
Go structs, this normalizes a `config.Value` according to the type of a Go struct
and returns the new, normalized `config.Value`.

This will be used to ensure that two config.Value trees are type-compatible
before we can merge them (i.e. instances from different files).

Warnings and errors during normalization are accumulated and returned as a
`diag.Diagnostics` structure. We can use this to surface warnings about unknown
fields, or errors about invalid types, in aggregate instead of one-by-one.