Rewrite the friendly log handler
It wasn't working because it deferred to the regular `slog.TextHandler` for the
`WithAttr` and `WithGroup` functions. Both of these functions don't mutate the
handler but return a new one. When the top level logger called one of these,
log records in that context used the standard handler instead of ours.

To implement tracking of attrs and groups, I followed the guide at
https://github.com/golang/example/blob/master/slog-handler-guide/README.md for
writing custom handlers.